+++ BambooTracker/command/order/set_pattern_to_order_command.hpp
@@ -7,7 +7,7 @@
 class SetPatternToOrderCommand : public AbstractCommand
 {
 public:
-	SetPatternToOrderCommand(std::weak_ptr<Module> mod, int songNum, int trackNum, int orderNum, int patternNum);
+	SetPatternToOrderCommand(std::weak_ptr<Module> mod, int songNum, int trackNum, int orderNum, int patternNum, bool secondEntry);
 	void redo() override;
 	void undo() override;
 	CommandId getID() const override;
@@ -16,11 +16,12 @@ public:
 	int getSong() const;
 	int getTrack() const;
 	int getOrder() const;
+	bool isSecondEntry() const;
 	int getPattern() const;
 
 private:
 	std::weak_ptr<Module> mod_;
 	int song_, track_, order_, pattern_;
 	int prevPattern_;
-	bool isComplete_;
+	bool isSecond_;
 };