+++ BambooTracker/command/pattern/set_effect_id_to_step_command.hpp
@@ -9,7 +9,7 @@ class SetEffectIDToStepCommand : public AbstractCommand
 {
 public:
 	SetEffectIDToStepCommand(std::weak_ptr<Module> mod, int songNum, int trackNum, int orderNum, int stepNum,
-							 int n, std::string id, bool fillValue00);
+							 int n, std::string id, bool fillValue00, bool secondEntry);
 	void redo() override;
 	void undo() override;
 	CommandId getID() const override;
@@ -20,11 +20,13 @@ public:
 	int getOrder() const;
 	int getStep() const;
 	int getN() const;
+	bool isSecondEntry() const;
 	std::string getEffectID() const;
 
 private:
 	std::weak_ptr<Module> mod_;
 	int song_, track_, order_, step_, n_;
 	std::string effID_, prevEffID_;
-	bool isComplete_, filledValue00_;
+	bool filledValue00_;
+	bool isSecond_;
 };