+++ BambooTracker/command/pattern/set_effect_value_to_step_command.hpp
@@ -8,7 +8,7 @@ class SetEffectValueToStepCommand : public AbstractCommand
 {
 public:
 	SetEffectValueToStepCommand(std::weak_ptr<Module> mod, int songNum, int trackNum,
-								int orderNum, int stepNum, int n, int value, bool isFMReversed);
+								int orderNum, int stepNum, int n, int value, bool isFMReversed, bool secondEntry);
 	void redo() override;
 	void undo() override;
 	CommandId getID() const override;
@@ -19,12 +19,13 @@ public:
 	int getOrder() const;
 	int getStep() const;
 	int getN() const;
+	bool isSecondEntry() const;
 	int getEffectValue() const;
 
 private:
 	std::weak_ptr<Module> mod_;
 	int song_, track_, order_, step_, n_;
 	int val_, prevVal_;
-	bool isComplete_;
 	bool isFMReserved_;
+	bool isSecond_;
 };