+++ BambooTracker/command/pattern/set_instrument_to_step_command.hpp
@@ -7,7 +7,7 @@
 class SetInstrumentToStepCommand : public AbstractCommand
 {
 public:
-	SetInstrumentToStepCommand(std::weak_ptr<Module> mod, int songNum, int trackNum, int orderNum, int stepNum, int instNum);
+	SetInstrumentToStepCommand(std::weak_ptr<Module> mod, int songNum, int trackNum, int orderNum, int stepNum, int instNum, bool secondEntry);
 	void redo() override;
 	void undo() override;
 	CommandId getID() const override;
@@ -17,11 +17,12 @@ public:
 	int getTrack() const;
 	int getOrder() const;
 	int getStep() const;
+	bool isSecondEntry() const;
 	int getInst() const;
 
 private:
 	std::weak_ptr<Module> mod_;
 	int song_, track_, order_, step_, inst_;
 	int prevInst_;
-	bool isComplete_;
+	bool isSecond_;
 };