+++ BambooTracker/command/pattern/set_volume_to_step_command.hpp
@@ -8,7 +8,7 @@ class SetVolumeToStepCommand : public AbstractCommand
 {
 public:
         SetVolumeToStepCommand(std::weak_ptr<Module> mod, int songNum, int trackNum,
-                               int orderNum, int stepNum, int volume, bool isFMReversed);
+							   int orderNum, int stepNum, int volume, bool isFMReversed, bool secondEntry);
 	void redo() override;
 	void undo() override;
 	CommandId getID() const override;
@@ -18,12 +18,13 @@ public:
 	int getTrack() const;
 	int getOrder() const;
 	int getStep() const;
+	bool isSecondEntry() const;
 	int getVol() const;
 
 private:
 	std::weak_ptr<Module> mod_;
 	int song_, track_, order_, step_, vol_;
 	int prevVol_;
-	bool isComplete_;
 	bool isFMReserved_;
+	bool isSecond_;
 };